+++ README.md
@@ -62,8 +62,7 @@ ReactDOM.render(<Pagination />, container);
 | defaultPageSize              | default items per page                                                                                                                                                                                                                    | Number                                                                                                                                               | 10                                                                                     |
 | pageSize                     | items per page                                                                                                                                                                                                                            | Number                                                                                                                                               | 10                                                                                     |
 | onChange                     | page change callback                                                                                                                                                                                                                      | Function(current, pageSize)                                                                                                                          | -                                                                                      |
-| showSizeChanger              | show pageSize changer                                                                                                                                                                                                                     | Bool                                                                                                                                                 | `false` when total less then `totalBoundaryShowSizeChanger`, `true` when otherwise     |
-| pageSizeChanger              | Specify `options` as sizeChanger selections. `showSearch` in the sizeChanger Select component allows searching the size, with the default value being false. `onChange` returns the size change value only through the callback function. | Object -> { options: ['10', '20', '50', '100'], showSearch: false, onChange: Function(size) }                                                        | { showSearch: false }                                                                  |
+| showSizeChanger              | show pageSize changer                                                                                                                                                                                                                     | Boolean or Object => { options: ['10', '20', '50', '100'], showSearch: false, onChange: Function(size) }                                                                                                                                                 |  `false` when the total is less than `totalBoundaryShowSizeChanger`, `true` otherwise, or specify `options` as sizeChanger selections. `showSearch` in the sizeChanger Select component allows searching the size, with the default value being false. `onChange` returns the size change value only through the callback function. The object should be formatted as: `{ options: ['10', '20', '50', '100'], showSearch: false, onChange: Function(size) }`. The value can be either `true/false` or `{ showSearch: false, options:['10', '20', '50', '100'], onChange: (size: number) => void }`.  |
 | totalBoundaryShowSizeChanger | when total larger than it, `showSizeChanger` will be true                                                                                                                                                                                 | number                                                                                                                                               | 50                                                                                     |
 | pageSizeOptions              | specify the sizeChanger selections                                                                                                                                                                                                        | Array<String>                                                                                                                                        | ['10', '20', '50', '100']                                                              |
 | onShowSizeChange             | pageSize change callback                                                                                                                                                                                                                  | Function(current, size)                                                                                                                              | -                                                                                      |
