--- docs/demo/pageSizeChanger.md
@@ -1,8 +1,0 @@
----
-title: pageSizeChanger
-nav:
-  title: Demo
-  path: /demo
----
-
-<code src="../examples/pageSizeChanger.tsx"></code>
