+++ docs/demo/showSizeChanger.md
@@ -1,0 +1,8 @@
+---
+title: showSizeChanger
+nav:
+  title: Demo
+  path: /demo
+---
+
+<code src="../examples/showSizeChanger.tsx"></code>
